refactor: Reorganize model classes and clean up unused imports
