+++ src/main/kotlin/com/example/demo/controller/AuthController.kt
@@ -1,9 +1,7 @@
 package com.example.demo.controller
 
 import com.example.demo.dto.LoginRequestDTO
-import com.example.demo.dto.UserResponseDTO
 import com.example.demo.dto.user.UserDTO
-import com.example.demo.exception.UserNotFoundException
 import com.example.demo.service.UserService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
