--- src/main/kotlin/com/example/demo/dto/UserInviteRequestDTO.kt
@@ -1,10 +1,0 @@
-package com.example.demo.dto
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class UserInviteRequestDTO (
-    val channelId: String,
-    val deviceIds: List<String>
-)
