+++ src/main/kotlin/com/example/demo/dto/channel/ChannelResponseDTO.kt
@@ -1,11 +1,10 @@
-package com.example.demo.dto
+package com.example.demo.dto.channel
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class CreateChannelRequest(
-    val name: String,
-    val deviceIds: List<String>,
-    val operatorIds: List<String>
+data class ChannelResponseDTO(
+    val channelName: String,
+    val devices: List<String>
 )
