+++ src/main/kotlin/com/example/demo/dto/channel/CreateChannelRequest.kt
@@ -1,15 +1,11 @@
-package com.example.demo.dto
+package com.example.demo.dto.channel
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
-import java.time.LocalDate
-import java.util.UUID
-
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class UserResponseDTO(
-    val id: UUID,
+data class CreateChannelRequest(
     val name: String,
-    val birthDate: LocalDate,
-    val phoneNumber: String
+    val deviceIds: List<String>,
+    val operatorIds: List<String>
 )
