+++ src/main/kotlin/com/example/demo/dto/device/DeviceStatusRequestDTO.kt
@@ -1,11 +1,12 @@
-package com.example.demo.dto
+package com.example.demo.dto.device
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 import java.util.*
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class DeviceUsageRequestDTO(
-    val userId: UUID,
-    val deviceId: UUID
+data class DeviceStatusRequestDTO(
+    val channelId: UUID,
+    val deviceId: UUID,
+    val deviceStatus: Boolean
 )
