+++ src/main/kotlin/com/example/demo/dto/device/DeviceUsageCreateDTO.kt
@@ -1,11 +1,11 @@
-package com.example.demo.dto
+package com.example.demo.dto.device
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class ChannelDeviceDTO(
-    val id: String,
-    val name: String,
-    val category: String?,
+data class DeviceUsageCreateDTO (
+    val  userId: String,
+    val  deviceId: String,
+    val  actionDescription: String
 )
