+++ src/main/kotlin/com/example/demo/dto/device/DeviceUsageRequestDTO.kt
@@ -1,11 +1,11 @@
-package com.example.demo.dto
+package com.example.demo.dto.device
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
+import java.util.*
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class DeviceUsageCreateDTO (
-    val  userId: String,
-    val  deviceId: String,
-    val  actionDescription: String
+data class DeviceUsageRequestDTO(
+    val userId: UUID,
+    val deviceId: UUID
 )
