+++ src/main/kotlin/com/example/demo/dto/device/DeviceUsageResponseDTO.kt
@@ -1,5 +1,7 @@
-package com.example.demo.dto
+package com.example.demo.dto.device
 
+import com.example.demo.dto.ActionDTO
+import com.example.demo.dto.channel.ChannelDTO
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
