+++ src/main/kotlin/com/example/demo/dto/user/UserInviteRequestDTO.kt
@@ -1,12 +1,10 @@
-package com.example.demo.dto
+package com.example.demo.dto.user
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
-import java.util.*
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class DeviceStatusRequestDTO(
-    val channelId: UUID,
-    val deviceId: UUID,
-    val deviceStatus: Boolean
+data class UserInviteRequestDTO (
+    val channelId: String,
+    val deviceIds: List<String>
 )
