+++ src/main/kotlin/com/example/demo/model/ChannelDevice.kt
@@ -2,6 +2,7 @@ package com.example.demo.model
 
 import com.fasterxml.jackson.annotation.JsonBackReference
 import com.fasterxml.jackson.annotation.JsonProperty
+
 import jakarta.persistence.*
 import java.util.UUID
